src/bridge/requirements.txt
boto3
google-cloud-pubsub
